+++ src/pyochain/collections/__init__.py
@@ -1,6 +1,7 @@
 """Additional collection types."""
 
+from ._counter import PyoCounter
 from ._deque import Deque
 from ._stable_set import StableSet
 
-__all__ = ["Deque", "StableSet"]
+__all__ = ["Deque", "PyoCounter", "StableSet"]